Split ApiMethods class into partials and add in GlobalCommands
