+++ Assets/Scripts/API/ApiMethods.cs
@@ -13,126 +13,32 @@
 // limitations under the License.
 
 using System;
-using System.Collections;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
-using Newtonsoft.Json;
-using SVGMeshUnity;
+using System.Net;
 using UnityEngine;
 
 namespace TiltBrush
 {
     // ReSharper disable once UnusedType.Global
-    public static class ApiMethods
+    public static partial class ApiMethods
     {
-
-        public static Quaternion LookAt(Vector3 sourcePoint, Vector3 destPoint)
-        {
-            Vector3 forwardVector = Vector3.Normalize(destPoint - sourcePoint);
-            float dot = Vector3.Dot(Vector3.forward, forwardVector);
-            if (Math.Abs(dot - (-1.0f)) < 0.000001f)
-            {
-                return new Quaternion(Vector3.up.x, Vector3.up.y, Vector3.up.z, 3.1415926535897932f);
-            }
-            if (Math.Abs(dot - (1.0f)) < 0.000001f)
-            {
-                return Quaternion.identity;
-            }
-
-            float rotAngle = (float)Math.Acos(dot);
-            Vector3 rotAxis = Vector3.Cross(Vector3.forward, forwardVector);
-            rotAxis = Vector3.Normalize(rotAxis);
-            return CreateFromAxisAngle(rotAxis, rotAngle);
-        }
-
-        public static Quaternion CreateFromAxisAngle(Vector3 axis, float angle)
-        {
-            float halfAngle = angle * .5f;
-            float s = (float)System.Math.Sin(halfAngle);
-            Quaternion q;
-            q.x = axis.x * s;
-            q.y = axis.y * s;
-            q.z = axis.z * s;
-            q.w = (float)System.Math.Cos(halfAngle);
-            return q;
-        }
-
-        private static void ChangeBrushBearing(float angle, Vector3 axis)
-        {
-            ApiManager.Instance.BrushRotation *= Quaternion.AngleAxis(angle, axis);
-        }
-
-        private static void ChangeSpectatorBearing(float angle, Vector3 axis)
-        {
-            var cam = SketchControlsScript.m_Instance.GetDropCampWidget();
-            cam.transform.rotation *= Quaternion.AngleAxis(angle, axis);
-        }
-
-        // private static void ChangeUserBearing(float angle, Vector3 axis)
+        
+        // Example of calling a command and recording an undo step
+        // [ApiEndpoint("foo", "")]
+        // public static void FooCommand()
         // {
-        //     Transform camTr = Camera.main.transform;
-        //     var rot = camTr.rotation;
-        //     rot *= Quaternion.AngleAxis(angle, axis);
-        //     camTr.rotation = rot;
+        //     FooCommand cmd = new FooCommand();
+        //     SketchMemoryScript.m_Instance.PerformAndRecordCommand(cmd);
         // }
-
-        [ApiEndpoint("draw.paths", "Draws a series of paths at the current brush position [[[x1,y1,z1],[x2,y2,z2], etc...]]. Does not move the brush position")]
-        public static void DrawPaths(string jsonString)
-        {
-            var origin = ApiManager.Instance.BrushPosition;
-            var paths = JsonConvert.DeserializeObject<List<List<List<float>>>>($"[{jsonString}]");
-            DrawStrokes.MultiPathsToStrokes(paths, origin);
-        }
-
-        [ApiEndpoint("draw.path", "Draws a path at the current brush position [x1,y1,z1],[x2,y2,z2], etc.... Does not move the brush position")]
-        public static void DrawPath(string jsonString)
-        {
-            var origin = ApiManager.Instance.BrushPosition;
-            var path = JsonConvert.DeserializeObject<List<List<float>>>($"[{jsonString}]");
-            DrawStrokes.SinglePathToStroke(path, origin);
-        }
-
-        [ApiEndpoint("draw.stroke", "Draws an exact brush stroke as recorded in another app")]
-        public static void DrawStroke(string jsonString)
-        {
-            var strokeData = JsonConvert.DeserializeObject<List<List<float>>>($"[{jsonString}]");
-            DrawStrokes.SinglePathToStroke(strokeData, Vector3.zero, rawStroke: true);
-        }
 
         [ApiEndpoint("listenfor.strokes", "Adds the url of an app that wants to receive the data for a stroke as each one is finished")]
         public static void AddListener(string url)
         {
             ApiManager.Instance.AddOutgoingCommandListener(new Uri(url));
         }
-
-        private static Vector3 rotatePointAroundPivot(Vector3 point, Vector3 pivot, Quaternion rot)
-        {
-            Vector3 dir = point - pivot;
-            dir = rot * dir;
-            point = dir + pivot;
-            return point;
-        }
-
-        [ApiEndpoint("draw.polygon", "Draws a polygon at the current brush position. Does not move the brush position")]
-        public static void DrawPolygon(int sides, float radius, float angle)
-        {
-            var path = new List<Vector3>();
-            for (float i = 0; i <= sides; i++)
-            {
-                var theta = Mathf.PI * (i / sides) * 2f;
-                theta += angle * Mathf.Deg2Rad;
-                var point = new Vector3(
-                    Mathf.Cos(theta),
-                    Mathf.Sin(theta),
-                    0
-                ) * radius;
-                point = ApiManager.Instance.BrushRotation * point;
-                path.Add(point);
-            }
-            DrawStrokes.PositionPathsToStroke(path, ApiManager.Instance.BrushPosition);
-        }
-
+        
         [ApiEndpoint("showfolder.scripts", "Opens the user's Scripts folder on the desktop")]
         public static void OpenUserScriptsFolder()
         {
@@ -157,128 +63,6 @@ namespace TiltBrush
                 );
                 System.Diagnostics.Process.Start("explorer.exe", "/select," + path);
             }
-        }
-
-        [ApiEndpoint("draw.text", "Draws the characters supplied at the current brush position")]
-        public static void Text(string text)
-        {
-            var origin = ApiManager.Instance.BrushPosition;
-            var font = Resources.Load<CHRFont>("arcade");
-            var textToStroke = new TextToStrokes(font);
-            var polyline2d = textToStroke.Build(text);
-            DrawStrokes.MultiPositionPathsToStrokes(polyline2d, null, null, origin);
-        }
-
-        [ApiEndpoint("draw.svg", "Draws the path supplied as an SVG Path string at the current brush position")]
-        public static void SvgPath(string svgPathString)
-        {
-            var origin = ApiManager.Instance.BrushPosition;
-            SVGData svgData = new SVGData();
-            svgData.Path(svgPathString);
-            SVGPolyline svgPolyline = new SVGPolyline();
-            svgPolyline.Fill(svgData);
-            DrawStrokes.MultiPath2dToStrokes(svgPolyline.Polyline, origin, 0.01f, true);
-        }
-
-        [ApiEndpoint("brush.type", "Changes the brush. brushType can either be the brush name or it's guid. brushes are listed in the localhost:40074/help screen")]
-        public static void Brush(string brushType)
-        {
-            BrushDescriptor brushDescriptor = null;
-            try
-            {
-                var guid = new Guid(brushType);
-                brushDescriptor = BrushCatalog.m_Instance.GetBrush(guid);
-            }
-            catch (FormatException e)
-            {
-            }
-
-            if (brushDescriptor == null)
-            {
-                brushType = brushType.ToLower().Trim().Replace(" ", "");
-                try
-                {
-                    brushDescriptor = BrushCatalog.m_Instance.AllBrushes
-                        .First(x => x.m_Description
-                            .Replace(" ", "")
-                            .Replace(".", "")
-                            .Replace("(", "")
-                            .Replace(")", "")
-                            .ToLower() == brushType);
-                }
-                catch (InvalidOperationException e)
-                {
-                    Debug.LogError($"No brush found called: {brushType}");
-                }
-            }
-
-            if (brushDescriptor != null)
-            {
-                PointerManager.m_Instance.SetBrushForAllPointers(brushDescriptor);
-            }
-            else
-            {
-                Debug.LogError($"No brush found with the name or guid: {brushType}");
-            }
-        }
-
-        [ApiEndpoint("color.add.hsv", "Adds the supplied values to the current color. Values are hue, saturation and value")]
-        public static void AddColorHSV(Vector3 hsv)
-        {
-            float h, s, v;
-            Color.RGBToHSV(App.BrushColor.CurrentColor, out h, out s, out v);
-            App.BrushColor.CurrentColor = Color.HSVToRGB(
-                (h + hsv.x) % 1f,
-                (s + hsv.y) % 1f,
-                (v + hsv.z) % 1f
-            );
-        }
-
-        [ApiEndpoint("color.add.rgb", "Adds the supplied values to the current color. Values are red green and blue")]
-        public static void AddColorRGB(Vector3 rgb)
-        {
-            App.BrushColor.CurrentColor += new Color(rgb.x, rgb.y, rgb.z);
-        }
-
-        [ApiEndpoint("color.set.rgb", "Sets the current color. Values are hue, saturation and value")]
-        public static void SetColorRGB(Vector3 rgb)
-        {
-            App.BrushColor.CurrentColor = new Color(rgb.x, rgb.y, rgb.z);
-        }
-
-        [ApiEndpoint("color.set.hsv", "Sets the current color. Values are red, green and blue")]
-        public static void SetColorHSV(Vector3 hsv)
-        {
-            App.BrushColor.CurrentColor = Color.HSVToRGB(hsv.x, hsv.y, hsv.z);
-        }
-
-        [ApiEndpoint("color.set.html", "Sets the current color. colorString can either be a hex value or a css color name.")]
-        public static void SetColorHTML(string color)
-        {
-            Color c;
-            color = color.ToLower();
-            if (CssColors.NamesToHex.ContainsKey(color)) color = CssColors.NamesToHex[color];
-            if (!color.StartsWith("#")) color = $"#{color}";
-            if (ColorUtility.TryParseHtmlString(color, out c))
-            {
-                App.BrushColor.CurrentColor = c;
-            }
-            else
-            {
-                Debug.LogError($"Invalid color: {color}");
-            }
-        }
-
-        [ApiEndpoint("brush.size.set", "Sets the current brush size")]
-        public static void BrushSizeSet(float size)
-        {
-            PointerManager.m_Instance.MainPointer.BrushSize01 = size;
-        }
-
-        [ApiEndpoint("brush.size.add", "Changes the current brush size by 'amount'")]
-        public static void BrushSizeAdd(float amount)
-        {
-            PointerManager.m_Instance.MainPointer.BrushSize01 += amount;
         }
 
         [ApiEndpoint("spectator.move.to", "Moves the spectator camera to the given position")]
@@ -318,19 +102,19 @@ namespace TiltBrush
         [ApiEndpoint("spectator.turn.y", "Rotates the spectator camera left or right.")]
         public static void SpectatorYaw(float angle)
         {
-            ChangeSpectatorBearing(angle, Vector3.up);
+            _ChangeSpectatorBearing(angle, Vector3.up);
         }
 
         [ApiEndpoint("spectator.turn.x", "Rotates the spectator camera up or down.")]
         public static void SpectatorPitch(float angle)
         {
-            ChangeSpectatorBearing(angle, Vector3.left);
+            _ChangeSpectatorBearing(angle, Vector3.left);
         }
 
         [ApiEndpoint("spectator.turn.z", "Tilts the angle of the spectator camera clockwise or anticlockwise.")]
         public static void SpectatorRoll(float angle)
         {
-            ChangeSpectatorBearing(angle, Vector3.forward);
+            _ChangeSpectatorBearing(angle, Vector3.forward);
         }
 
         // [ApiEndpoint("user.turn.y", "Rotates the user camera left or right.")]
@@ -338,13 +122,13 @@ namespace TiltBrush
         // {
         //     ChangeUserBearing(angle, Vector3.up);
         // }
-        //
+        
         // [ApiEndpoint("user.turn.x", "Rotates the user camera up or down. (monoscopic mode only)")]
         // public static void UserPitch(float angle)
         // {
         //     ChangeUserBearing(angle, Vector3.left);
         // }
-        //
+        
         // [ApiEndpoint("user.turn.z", "Tilts the angle of the user camera clockwise or anticlockwise. (monoscopic mode only)")]
         // public static void UserRoll(float angle)
         // {
@@ -373,7 +157,7 @@ namespace TiltBrush
         public static void SpectatorLookAt(Vector3 position)
         {
             var cam = SketchControlsScript.m_Instance.GetDropCampWidget();
-            Quaternion qNewRotation = LookAt(cam.transform.position, position);
+            Quaternion qNewRotation = _LookAt(cam.transform.position, position);
             cam.transform.rotation = qNewRotation;
         }
 
@@ -407,59 +191,16 @@ namespace TiltBrush
             }
         }
 
-        public static void SpectatorShowHide(string thing, bool state)
-        {
-            // Friendly names to layer names
-            string layerName = null;
-            switch (thing.Trim().ToLower())
-            {
-                case "widgets":
-                    layerName = "GrabWidgets";
-                    break;
-                case "strokes":
-                    layerName = "MainCanvas";
-                    break;
-                case "selection":
-                    layerName = "SelectionCanvas";
-                    break;
-                case "headset":
-                    layerName = "HeadMesh";
-                    break;
-                case "panels":
-                    layerName = "Panels";
-                    break;
-                case "ui":
-                    layerName = "UI";
-                    break;
-                case "usertools":
-                    layerName = "UserTools";
-                    break;
-            }
-
-            if (layerName == null) return;
-
-            int mask = 1 << LayerMask.NameToLayer(layerName);
-            Camera cam = SketchControlsScript.m_Instance.GetDropCampWidget().GetComponentInChildren<Camera>();
-            if (state)
-            {
-                cam.cullingMask |= mask;
-            }
-            else
-            {
-                cam.cullingMask = ~(~cam.cullingMask | mask);
-            }
-        }
-
         [ApiEndpoint("spectator.show", "Unhides the chosen type of elements from the spectator camera (widgets, strokes, selection, headset, panels, ui")]
         public static void SpectatorShow(string thing)
         {
-            SpectatorShowHide(thing, true);
+            _SpectatorShowHide(thing, true);
         }
 
         [ApiEndpoint("spectator.hide", "Hides the chosen type of elements from the spectator camera (widgets, strokes, selection, headset, panels, ui")]
         public static void SpectatorHide(string thing)
         {
-            SpectatorShowHide(thing, false);
+            _SpectatorShowHide(thing, false);
         }
 
         [ApiEndpoint("brush.move.to", "Moves the brush to the given coordinates")]
@@ -500,19 +241,19 @@ namespace TiltBrush
         [ApiEndpoint("brush.turn.y", "Changes the brush direction to the left or right. Angle is measured in degrees")]
         public static void BrushYaw(float angle)
         {
-            ChangeBrushBearing(angle, Vector3.up);
+            _ChangeBrushBearing(angle, Vector3.up);
         }
 
         [ApiEndpoint("brush.turn.x", "Changes the brush direction up or down. Angle is measured in degrees")]
         public static void BrushPitch(float angle)
         {
-            ChangeBrushBearing(angle, Vector3.left);
+            _ChangeBrushBearing(angle, Vector3.left);
         }
 
         [ApiEndpoint("brush.turn.z", "Rotates the brush clockwise or anticlockwise. Angle is measured in degrees")]
         public static void BrushRoll(float angle)
         {
-            ChangeBrushBearing(angle, Vector3.forward);
+            _ChangeBrushBearing(angle, Vector3.forward);
         }
 
         [ApiEndpoint("brush.look.at", "Changes the brush direction to look at the specified point")]
@@ -592,257 +333,132 @@ namespace TiltBrush
             Debug.Log($"Brush rotation: {ApiManager.Instance.BrushRotation.eulerAngles}");
         }
 
-        [ApiEndpoint("stroke.delete", "Delete strokes by their index. If index is 0 the most recent stroke is deleted. -1 etc steps back in time")]
-        public static void DeleteStroke(int index)
+        private static ReferenceImage _LoadReferenceImage(string location)
         {
-            var stroke = SketchMemoryScript.m_Instance.GetStrokeAtIndex(index);
-            SketchMemoryScript.m_Instance.RemoveMemoryObject(stroke);
-            stroke.Uncreate();
-        }
-
-        [ApiEndpoint("stroke.select", "Selects a stroke by it's index. 0 is the most recent stroke, -1 is second to last, 1 is the first.")]
-        public static void SelectStroke(int index)
-        {
-            var stroke = SketchMemoryScript.m_Instance.GetStrokeAtIndex(index);
-            SelectionManager.m_Instance.SelectStrokes(new List<Stroke> { stroke });
-        }
-
-        [ApiEndpoint("strokes.select", "Select multiple strokes by their index. 0 is the most recent stroke, -1 is second to last, 1 is the first.")]
-        public static void SelectStrokes(int start, int end)
-        {
-            var strokes = SketchMemoryScript.GetStrokesBetween(start, end);
-            SelectionManager.m_Instance.SelectStrokes(strokes);
-        }
-
-        [ApiEndpoint("selection.recolor", "Recolors the currently selected strokes")]
-        public static void RecolorSelection()
-        {
-            foreach (Stroke stroke in SelectionManager.m_Instance.SelectedStrokes)
+            location = Path.Combine(App.MediaLibraryPath(), "Images", location);
+            var image = new ReferenceImage(location);
+            bool result;
+            int timeout = 0;
+            do
             {
-                SketchMemoryScript.m_Instance.MemorizeStrokeRepaint(stroke, true, false, false);
-            }
+                result = image.RequestLoad();
+            } while (result==false && timeout++ < 100000);
+            return image;
         }
-
-        [ApiEndpoint("selection.rebrush", "Rebrushes the currently selected strokes")]
-        public static void RebrushSelection()
+        
+        [ApiEndpoint("image.import", "Imports an image given a url or a filename in Media Library\\Images")]
+        public static void ImportImage(string location)
         {
-            foreach (Stroke stroke in SelectionManager.m_Instance.SelectedStrokes)
-            {
-                SketchMemoryScript.m_Instance.MemorizeStrokeRepaint(stroke, false, true, false);
-            }
-        }
-
-        [ApiEndpoint("selection.resize", "Changes the brush size the currently selected strokes")]
-        public static void ResizeSelection()
-        {
-            foreach (Stroke stroke in SelectionManager.m_Instance.SelectedStrokes)
-            {
-                SketchMemoryScript.m_Instance.MemorizeStrokeRepaint(stroke, false, false, true);
-            }
-        }
-
-        [ApiEndpoint("selection.trim", "Removes a number of points from the currently selected strokes")]
-        public static void TrimSelection(int count)
-        {
-            foreach (Stroke stroke in SelectionManager.m_Instance.SelectedStrokes)
-            {
-                int newCount = Mathf.Max(0, stroke.m_ControlPoints.Length - count);
-                if (newCount > 0)
-                {
-                    Array.Resize(ref stroke.m_ControlPoints, newCount);
-                    stroke.Recreate(null, stroke.Canvas);
-                    SketchMemoryScript.m_Instance.MemorizeStrokeRepaint(stroke, false, false, true);
-                }
-                else
-                {
-                    SketchMemoryScript.m_Instance.RemoveMemoryObject(stroke);
-                    stroke.Uncreate();
-                }
-            }
-        }
-
-        private static Vector3 QuantizePosition(Vector3 pos, Vector3 grid)
-        {
-            float round(float val, float res) { return Mathf.Round(val / res) * res; }
-            return new Vector3(round(pos.x, grid.x), round(pos.y, grid.y), round(pos.z, grid.z));
-        }
-
-        private static void ModifyControlPoints(Func<Vector3, Vector3> func)
-        {
-            foreach (Stroke stroke in SelectionManager.m_Instance.SelectedStrokes)
-            {
-                var newCPs = new List<PointerManager.ControlPoint>();
-                for (var i = 0; i < stroke.m_ControlPoints.Length; i++)
-                {
-                    var cp = stroke.m_ControlPoints[i];
-                    cp.m_Pos = func(cp.m_Pos);
-                    // Only add a control point if it's pos is different to it's predecessor
-                    if (i == 0 || (i > 0 && cp.m_Pos != stroke.m_ControlPoints[i - 1].m_Pos))
-                    {
-                        newCPs.Add(cp);
-                    }
-                }
-                stroke.m_ControlPoints = newCPs.ToArray();
-                stroke.Uncreate();
-                stroke.Recreate(null, stroke.Canvas);
-            }
-        }
-
-        [ApiEndpoint("selection.points.addnoise", "Moves the position of all control points in the selection using a noise function")]
-        public static void PerlinNoiseSelection(string axis, Vector3 scale)
-        {
-            Enum.TryParse(axis.ToUpper(), out Axis _axis);
-            Func<Vector3, Vector3> quantize = pos => PerlinNoisePosition(pos, scale, _axis);
-            ModifyControlPoints(quantize);
-        }
-
-        private enum Axis { X, Y, Z }
-
-        private static Vector3 PerlinNoisePosition(Vector3 pos, Vector3 scale, Axis axis)
-        {
-            pos = new Vector3(pos.x / scale.x, pos.y / scale.y, pos.z / scale.z);
-            switch (axis)
-            {
-                case Axis.X:
-                    pos.x += Mathf.PerlinNoise(pos.y, pos.z) * scale.x;
-                    break;
-                case Axis.Y:
-                    pos.y += Mathf.PerlinNoise(pos.x, pos.z) * scale.y;
-                    break;
-                case Axis.Z:
-                    pos.z += Mathf.PerlinNoise(pos.x, pos.y) * scale.z;
-                    break;
-            }
-            return new Vector3(pos.x * scale.x, pos.y * scale.y, pos.z * scale.z);
-        }
-
-        [ApiEndpoint("selection.points.quantize", "Snaps all the points in selected strokes to a grid (buggy)")]
-        public static void QuantizeSelection(Vector3 grid)
-        {
-            Func<Vector3, Vector3> quantize = pos => QuantizePosition(pos, grid);
-            ModifyControlPoints(quantize);
-        }
-
-        [ApiEndpoint("stroke.join", "Joins a stroke with the previous one")]
-        public static void JoinStroke()
-        {
-            var stroke1 = SketchMemoryScript.m_Instance.GetStrokeAtIndex(0);
-            var stroke2 = SketchMemoryScript.m_Instance.GetStrokeAtIndex(-1);
-            stroke2.m_ControlPoints = stroke2.m_ControlPoints.Concat(stroke1.m_ControlPoints).ToArray();
-            stroke2.Uncreate();
-            stroke2.m_ControlPointsToDrop = Enumerable.Repeat(false, stroke2.m_ControlPoints.Length).ToArray();
-            stroke2.Recreate(null, stroke2.Canvas);
-            DeleteStroke(0);
-        }
-
-        [ApiEndpoint("strokes.join", "Joins all strokes between the two indices (inclusive)")]
-        public static void JoinStrokes(int start, int end)
-        {
-            var strokesToJoin = SketchMemoryScript.GetStrokesBetween(start, end);
-            var firstStroke = strokesToJoin[0];
-            firstStroke.m_ControlPoints = strokesToJoin.SelectMany(x => x.m_ControlPoints).ToArray();
-            for (int i = 1; i < strokesToJoin.Count; i++)
-            {
-                var stroke = strokesToJoin[i];
-                SketchMemoryScript.m_Instance.RemoveMemoryObject(stroke);
-                stroke.DestroyStroke();
-            }
-            firstStroke.Uncreate();
-            firstStroke.m_ControlPointsToDrop = Enumerable.Repeat(false, firstStroke.m_ControlPoints.Length).ToArray();
-            firstStroke.Recreate(null, firstStroke.Canvas);
-        }
-
-        [ApiEndpoint("stroke.add", "Adds a point at the current brush position to the specified stroke")]
-        public static void AddPointToStroke(int index)
-        {
-            var stroke = SketchMemoryScript.m_Instance.GetStrokeAtIndex(index);
-            var prevCP = stroke.m_ControlPoints[stroke.m_ControlPoints.Length - 1];
-            Array.Resize(ref stroke.m_ControlPoints, stroke.m_ControlPoints.Length + 1);
-
-            PointerManager.ControlPoint cp = new PointerManager.ControlPoint
-            {
-                m_Pos = ApiManager.Instance.BrushPosition,
-                m_Orient = ApiManager.Instance.BrushRotation,
-                m_Pressure = prevCP.m_Pressure,
-                m_TimestampMs = prevCP.m_TimestampMs
-            };
-
-            stroke.m_ControlPoints[stroke.m_ControlPoints.Length - 1] = cp;
-            stroke.Uncreate();
-            stroke.m_ControlPointsToDrop = Enumerable.Repeat(false, stroke.m_ControlPoints.Length).ToArray();
-            stroke.Recreate(null, stroke.Canvas);
-        }
-
-
-
-        [ApiEndpoint("import.model", "Imports a model from your Open Brush\\Media Library\\Models folder")]
-        public static void ImportModel(string path)
-        {
-            path = Path.Combine(App.MediaLibraryPath(), "Models", path);
-            var model = new Model(Model.Location.File(path));
-            model.LoadModel();
-
+            var image = _LoadReferenceImage(location);
             var tr = new TrTransform();
             tr.translation = ApiManager.Instance.BrushPosition;
             tr.rotation = ApiManager.Instance.BrushRotation;
-            CreateWidgetCommand createCommand = new CreateWidgetCommand(
-                WidgetManager.m_Instance.ModelWidgetPrefab, tr);
-            SketchMemoryScript.m_Instance.PerformAndRecordCommand(createCommand);
-            ModelWidget modelWidget = createCommand.Widget as ModelWidget;
-            modelWidget.Model = model;
-            modelWidget.Show(true);
-            createCommand.SetWidgetCost(modelWidget.GetTiltMeterCost());
+            var cmd = new CreateWidgetCommand(WidgetManager.m_Instance.ImageWidgetPrefab, tr);
+
+            SketchMemoryScript.m_Instance.PerformAndRecordCommand(cmd);
+            var imageWidget = cmd.Widget as ImageWidget;
+            if (imageWidget != null)
+            {
+                imageWidget.ReferenceImage = image;
+                imageWidget.Show(true);
+                cmd.SetWidgetCost(imageWidget.GetTiltMeterCost());
+            }
 
             WidgetManager.m_Instance.WidgetsDormant = false;
             SketchControlsScript.m_Instance.EatGazeObjectInput();
             SelectionManager.m_Instance.RemoveFromSelection(false);
-
         }
 
-        [ApiEndpoint("import.polymodel", "Imports a model from it's Google Poly ID")]
-        public static void ImportPolyModel(string assetId)
+        [ApiEndpoint("environment.type", "Sets the current environment")]
+        public static void SetEnvironment(string name)
         {
-            ApiManager.Instance.LoadPolyModel(assetId);
+            Environment env = EnvironmentCatalog.m_Instance.AllEnvironments.First(x => x.name == name);
+            SceneSettings.m_Instance.SetDesiredPreset(env, false, true);
+        }
+        
+        [ApiEndpoint("layer.add", "Adds a new layer")]
+        public static void AddLayer()
+        {
+            AddLayerCommand cmd = new AddLayerCommand(makeActive: true);
+            SketchMemoryScript.m_Instance.PerformAndRecordCommand(cmd);
+        }
+        
+        [ApiEndpoint("layer.delete", "Deletes a layer")]
+        public static void DeleteLayer(int layer)
+        {
+            DeleteLayerCommand cmd = new DeleteLayerCommand(layer);
+            SketchMemoryScript.m_Instance.PerformAndRecordCommand(cmd);
+        }
+        
+        [ApiEndpoint("layer.squash", "Move all brush strokes from one layer to another and removes the empty layer")]
+        public static void SquashLayer(int squashedLayer, int destinationLayer)
+        {
+            Debug.Log($"squashedLayer {squashedLayer} destinationLayer {destinationLayer}");
+            SquashLayerCommand cmd = new SquashLayerCommand(squashedLayer, destinationLayer);
+            SketchMemoryScript.m_Instance.PerformAndRecordCommand(cmd);
+        }
+        
+        [ApiEndpoint("layer.activate", "Make a layer active")]
+        public static void ActivateLayer(int layer)
+        {
+            ActivateLayerCommand cmd = new ActivateLayerCommand(App.Scene.GetCanvasByLayerIndex(layer));
+            SketchMemoryScript.m_Instance.PerformAndRecordCommand(cmd);
+        }
+        
+        [ApiEndpoint("layer.show", "Make a layer visible")]
+        public static void ShowLayer(int layer)
+        {
+            App.Scene.ShowLayer(layer);
+        }
+        
+        [ApiEndpoint("layer.hide", "Hide a layer")]
+        public static void HideLayer(int layer)
+        {
+            App.Scene.HideLayer(layer);
+        }
+        
+        [ApiEndpoint("layer.toggle", "Toggles a layer between visible and hidden")]
+        public static void ToggleLayer(int layer)
+        {
+            App.Scene.ToggleLayerVisibility(layer);
         }
 
-        [ApiEndpoint("import.webmodel", "Imports a model from a public url")]
-        public static void ImportModelFromUrl(string url)
+        [ApiEndpoint("model.select", "Selects a widget by index.")]
+        private static void SelectModel(int index)
         {
-            Uri uri = new Uri(url);
-            ApiManager.Instance.LoadPolyModel(uri);
+            SelectWidget(_GetActiveModel(index));
         }
-
-        [ApiEndpoint("import.image", "Imports an image  from your Open Brush\\Media Library\\Images folder")]
-        public static void ImportImage(string path)
+        
+        private static void SelectWidget(GrabWidget widget)
         {
-            path = Path.Combine(App.MediaLibraryPath(), "Images", path);
-            var image = new ReferenceImage(path);
-            image.RequestLoad(allowMainThread: true);
-            var tr = new TrTransform();
-            tr.translation = ApiManager.Instance.BrushPosition;
-            tr.rotation = ApiManager.Instance.BrushRotation;
-            CreateWidgetCommand createCommand = new CreateWidgetCommand(
-            WidgetManager.m_Instance.ImageWidgetPrefab, tr);
-            SketchMemoryScript.m_Instance.PerformAndRecordCommand(createCommand);
-            ImageWidget imageWidget = createCommand.Widget as ImageWidget;
-            imageWidget.ReferenceImage = image;
-            imageWidget.Show(true);
-            createCommand.SetWidgetCost(imageWidget.GetTiltMeterCost());
-
-            WidgetManager.m_Instance.WidgetsDormant = false;
-            SketchControlsScript.m_Instance.EatGazeObjectInput();
-            SelectionManager.m_Instance.RemoveFromSelection(false);
-
+            SelectionManager.m_Instance.SelectWidget(widget);
+        }
+        
+        [ApiEndpoint("model.position", "Move a model to the given coordinates")]
+        public static void PositionModel(int index, Vector3 position)
+        {
+            _PositionWidget(_GetActiveModel(index), position);
+        }
+        
+        [ApiEndpoint("image.position", "Move an image to the given coordinates")]
+        public static void PositionImage(int index, Vector3 position)
+        {
+            _PositionWidget(_GetActiveImage(index), position);
         }
 
         // WIP
-        // [ApiEndpoint("import.video", "Imports a video from your Open Brush\\Media Library\\Video folder")]
-        // public static void ImportVideo(string path)
+        // [ApiEndpoint("video.import", "Imports a video given a url or a filename in Media Library\\Videos")]
+        // public static void ImportVideo(string location)
         // {
-        //     path = Path.Combine("Videos", path);
+        //     if (location.StartsWith("http://") || location.StartsWith("https://"));
+        //     {
+        //         location = DownloadMediaFileFromUrl(location, "Videos");
+        //     }
+        //     location = DownloadMediaFileFromUrl(location, "Videos");
+        //
+        //     location = Path.Combine("Videos", location);
         //     var video = new TiltVideo();
-        //     video.FilePath = path;
-        //     Debug.Log($"FilePath: {video.FilePath}");
+        //     video.FilePath = location;
         //     VideoWidget.FromTiltVideo(video);
         //     // var tr = new TrTransform();
         //     // tr.translation = ApiManager.Instance.BrushPosition;
@@ -856,248 +472,42 @@ namespace TiltBrush
         //     // WidgetManager.m_Instance.WidgetsDormant = false;
         //     // SketchControlsScript.m_Instance.EatGazeObjectInput();
         //     // SelectionManager.m_Instance.RemoveFromSelection(false);
-        //
         // }
 
-        // Tools.
-
-        [ApiEndpoint("tool.sketchsurface", "Activates the SketchSurface")]
-        public static void ActivateSketchSurface()
+        [ApiEndpoint("guide.add", "Adds a guide to the scene")]
+        public static void AddGuide(string type)
         {
-            SketchSurfacePanel.m_Instance.EnableSpecificTool(BaseTool.ToolType.SketchSurface);
-        }
+            StencilType stencilType;
 
-        [ApiEndpoint("tool.selection", "Activates the Selection Tool")]
-        public static void ActivateSelection()
-        {
-            SketchSurfacePanel.m_Instance.EnableSpecificTool(BaseTool.ToolType.Selection);
-        }
-
-        [ApiEndpoint("tool.colorpicker", "Activates the Color Picker")]
-        public static void ActivateColorPicker()
-        {
-            SketchSurfacePanel.m_Instance.EnableSpecificTool(BaseTool.ToolType.ColorPicker);
-        }
-
-        [ApiEndpoint("tool.brushpicker", "Activates the Brush Picker")]
-        public static void ActivateBrushPicker()
-        {
-            SketchSurfacePanel.m_Instance.EnableSpecificTool(BaseTool.ToolType.BrushPicker);
-        }
-
-        [ApiEndpoint("tool.brushandcolorpicker", "Activates the Brush And Color Picker")]
-        public static void ActivateBrushAndColorPicker()
-        {
-            SketchSurfacePanel.m_Instance.EnableSpecificTool(BaseTool.ToolType.BrushAndColorPicker);
-        }
-
-        [ApiEndpoint("tool.sketchorigin", "Activates the SketchOrigin Tool")]
-        public static void ActivateSketchOrigin()
-        {
-            SketchSurfacePanel.m_Instance.EnableSpecificTool(BaseTool.ToolType.SketchOrigin);
-        }
-
-        [ApiEndpoint("tool.autogif", "Activates the AutoGif Tool")]
-        public static void ActivateAutoGif()
-        {
-            SketchSurfacePanel.m_Instance.EnableSpecificTool(BaseTool.ToolType.AutoGif);
-        }
-
-        [ApiEndpoint("tool.canvas", "Activates the Canvas Tool")]
-        public static void ActivateCanvasTool()
-        {
-            SketchSurfacePanel.m_Instance.EnableSpecificTool(BaseTool.ToolType.CanvasTool);
-        }
-
-        [ApiEndpoint("tool.transform", "Activates the Transform Tool")]
-        public static void ActivateTransformTool()
-        {
-            SketchSurfacePanel.m_Instance.EnableSpecificTool(BaseTool.ToolType.TransformTool);
-        }
-
-        [ApiEndpoint("tool.stamp", "Activates the Stamp Tool")]
-        public static void ActivateStampTool()
-        {
-            SketchSurfacePanel.m_Instance.EnableSpecificTool(BaseTool.ToolType.StampTool);
-        }
-
-        [ApiEndpoint("tool.freepaint", "Activates the FreePaint Tool")]
-        public static void ActivateFreePaintTool()
-        {
-            SketchSurfacePanel.m_Instance.EnableSpecificTool(BaseTool.ToolType.FreePaintTool);
-        }
-
-        [ApiEndpoint("tool.eraser", "Activates the Eraser Tool")]
-        public static void ActivateEraserTool()
-        {
-            SketchSurfacePanel.m_Instance.EnableSpecificTool(BaseTool.ToolType.EraserTool);
-        }
-
-        [ApiEndpoint("tool.screenshot", "Activates the Screenshot Tool")]
-        public static void ActivateScreenshotTool()
-        {
-            SketchSurfacePanel.m_Instance.EnableSpecificTool(BaseTool.ToolType.ScreenshotTool);
-        }
-
-        [ApiEndpoint("tool.dropper", "Activates the Dropper Tool")]
-        public static void ActivateDropperTool()
-        {
-            SketchSurfacePanel.m_Instance.EnableSpecificTool(BaseTool.ToolType.DropperTool);
-        }
-
-        [ApiEndpoint("tool.saveicon", "Activates the SaveIcon Tool")]
-        public static void ActivateSaveIconTool()
-        {
-            SketchSurfacePanel.m_Instance.EnableSpecificTool(BaseTool.ToolType.SaveIconTool);
-        }
-
-        [ApiEndpoint("tool.threedofviewing", "Activates the ThreeDofViewing Tool")]
-        public static void ActivateThreeDofViewingTool()
-        {
-            SketchSurfacePanel.m_Instance.EnableSpecificTool(BaseTool.ToolType.ThreeDofViewingTool);
-        }
-
-        [ApiEndpoint("tool.multicam", "Activates the MultiCam Tool")]
-        public static void ActivateMultiCamTool()
-        {
-            SketchSurfacePanel.m_Instance.EnableSpecificTool(BaseTool.ToolType.MultiCamTool);
-        }
-
-        [ApiEndpoint("tool.teleport", "Activates the Teleport Tool")]
-        public static void ActivateTeleportTool()
-        {
-            SketchSurfacePanel.m_Instance.EnableSpecificTool(BaseTool.ToolType.TeleportTool);
-        }
-
-        [ApiEndpoint("tool.repaint", "Activates the Repaint Tool")]
-        public static void ActivateRepaintTool()
-        {
-            SketchSurfacePanel.m_Instance.EnableSpecificTool(BaseTool.ToolType.RepaintTool);
-        }
-
-        [ApiEndpoint("tool.recolor", "Activates the Recolor Tool")]
-        public static void ActivateRecolorTool()
-        {
-            SketchSurfacePanel.m_Instance.EnableSpecificTool(BaseTool.ToolType.RecolorTool);
-        }
-
-        [ApiEndpoint("tool.rebrush", "Activates the Rebrush Tool")]
-        public static void ActivateRebrushTool()
-        {
-            SketchSurfacePanel.m_Instance.EnableSpecificTool(BaseTool.ToolType.RebrushTool);
-        }
-
-        [ApiEndpoint("tool.selection", "Activates the Selection Tool")]
-        public static void ActivateSelectionTool()
-        {
-            SketchSurfacePanel.m_Instance.EnableSpecificTool(BaseTool.ToolType.SelectionTool);
-        }
-
-        [ApiEndpoint("tool.pin", "Activates the Pin Tool")]
-        public static void ActivatePinTool()
-        {
-            SketchSurfacePanel.m_Instance.EnableSpecificTool(BaseTool.ToolType.PinTool);
-        }
-
-        // [ApiEndpoint("tool.empty", "Activates the Empty Tool")]
-        // public static void ActivateEmptyTool()
-        // {
-        //     SketchSurfacePanel.m_Instance.EnableSpecificTool(BaseTool.ToolType.EmptyTool);
-        // }
-
-        [ApiEndpoint("tool.camerapath", "Activates the CameraPath Tool")]
-        public static void ActivateCameraPathTool()
-        {
-            SketchSurfacePanel.m_Instance.EnableSpecificTool(BaseTool.ToolType.CameraPathTool);
-        }
-
-        [ApiEndpoint("tool.fly", "Activates the Fly Tool")]
-        public static void ActivateFlyTool()
-        {
-            SketchSurfacePanel.m_Instance.EnableSpecificTool(BaseTool.ToolType.FlyTool);
-        }
-
-        [ApiEndpoint("environment.type", "Sets the current environment")]
-        public static void SetEnvironment(string name)
-        {
-            Environment env = EnvironmentCatalog.m_Instance.AllEnvironments.First(x => x.name == name);
-            SceneSettings.m_Instance.SetDesiredPreset(env, false, true);
-        }
-
-        [ApiEndpoint("draw.camerapath", "Draws along a camera path with the current brush settings")]
-        public static void DrawCameraPath(int index)
-        {
-            CameraPathWidget widget = WidgetManager.m_Instance.CameraPathWidgets.ElementAt(index).WidgetScript;
-            CameraPath path = widget.Path;
-            var positions = new List<Vector3>();
-            var rotations = new List<Quaternion>();
-            for (float t = 0; t < path.Segments.Count; t += .1f)
+            switch (type)
             {
-                positions.Add(path.GetPosition(new PathT(t)));
-                rotations.Add(path.GetRotation(new PathT(t)));
+                case "cube":
+                    stencilType = StencilType.Cube;
+                    break;
+                case "sphere":
+                    stencilType = StencilType.Sphere;
+                    break;
+                case "capsure":
+                    stencilType = StencilType.Capsule;
+                    break;
+                case "cone":
+                    stencilType = StencilType.Cone;
+                    break;
+                case "ellipsoid":
+                    stencilType = StencilType.Ellipsoid;
+                    break;
+                default:
+                    stencilType = StencilType.Sphere;
+                    break;
             }
-            DrawStrokes.MultiPositionPathsToStrokes(
-                new List<List<Vector3>> { positions },
-                new List<List<Quaternion>> { rotations },
-                null,
-                Vector3.zero
-            );
+
+            var tr = _CurrentTransform();
+            CreateWidgetCommand createCommand = new CreateWidgetCommand(
+                WidgetManager.m_Instance.GetStencilPrefab(stencilType), tr);
+            SketchMemoryScript.m_Instance.PerformAndRecordCommand(createCommand);
         }
 
-        [ApiEndpoint("layer.add", "")]
-        public static void AddLayer()
-        {
-            AddLayerCommand cmd = new AddLayerCommand(makeActive: true);
-            SketchMemoryScript.m_Instance.PerformAndRecordCommand(cmd);
-        }
 
-        [ApiEndpoint("layer.delete", "")]
-        public static void DeleteLayer(int layer)
-        {
-            DeleteLayerCommand cmd = new DeleteLayerCommand(layer);
-            SketchMemoryScript.m_Instance.PerformAndRecordCommand(cmd);
-        }
-
-        [ApiEndpoint("layer.squash", "")]
-        public static void SquashLayer(int squashedLayer, int destinationLayer)
-        {
-            Debug.Log($"squashedLayer {squashedLayer} destinationLayer {destinationLayer}");
-            SquashLayerCommand cmd = new SquashLayerCommand(squashedLayer, destinationLayer);
-            SketchMemoryScript.m_Instance.PerformAndRecordCommand(cmd);
-        }
-
-        [ApiEndpoint("layer.activate", "")]
-        public static void ActivateLayer(int layer)
-        {
-            ActivateLayerCommand cmd = new ActivateLayerCommand(App.Scene.GetCanvasByLayerIndex(layer));
-            SketchMemoryScript.m_Instance.PerformAndRecordCommand(cmd);
-        }
-
-        [ApiEndpoint("layer.show", "")]
-        public static void ShowLayer(int layer)
-        {
-            App.Scene.ShowLayer(layer);
-        }
-
-        [ApiEndpoint("layer.hide", "")]
-        public static void HideLayer(int layer)
-        {
-            App.Scene.HideLayer(layer);
-        }
-
-        [ApiEndpoint("layer.toggle", "")]
-        public static void ToggleLayer(int layer)
-        {
-            App.Scene.ToggleLayerVisibility(layer);
-        }
-
-        // Example of calling a command and recording an undo step
-        // [ApiEndpoint("foo", "")]
-        // public static void FooCommand()
-        // {
-        //     FooCommand cmd = new FooCommand();
-        //     SketchMemoryScript.m_Instance.PerformAndRecordCommand(cmd);
-        // }
 
     }
 }
